# apps/logistics-backend/docs/swagger.yaml
basePath: /api
definitions:
...
  driver.CreateDriverRequest:
    properties:
      current_location:
        type: string
      email:
        type: string
      full_name:
        type: string
      vehicle_info:
        type: string
    required:
    - current_location
    - email
    - full_name
    - vehicle_info
    type: object
  driver.Driver:
    properties:
      available:
        type: boolean
      created_at:
        type: string
      current_location:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      vehicle_info:
        type: string
    type: object
  driver.UpdateDriverRequest:
    properties:
      column:
        type: string
      value: {}
    required:
    - column
    - value
    type: object
  feedback.CreateFeedbackRequest:
    properties:
      comments:
        type: string
      customer_id:
        type: string
      order_id:
        type: string
      rating:
        type: integer
    type: object
  feedback.Feedback:
    properties:
      comments:
        type: string
      customer_id:
        type: string
      id:
        type: string
      order_id:
        type: string
      rating:
        type: integer
      submitted_at:
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      detail:
        description: optional internal error
        example: validation failed on field 'email'
        type: string
      error:
        description: user-friendly message
        example: Invalid request
        type: string
    type: object
...
host: 192.168.100.18:8000
info:
  contact: {}
  description: This is the API for logistics operations.
  title: Logistics API
  version: "1.0"
paths:
...
  /drivers/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently deletes a driver by their ID
      parameters:
      - description: Driver ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Driver deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid driver ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - JWT: []
      summary: Delete a driver
      tags:
      - drivers
  /drivers/{id}/profile:
    patch:
      consumes:
      - application/json
      description: Updates the vehicle information and current location of a driver
      parameters:
      - description: Driver ID
        in: path
        name: driver_id
        required: true
        type: string
      - description: Driver profile fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/driver.UpdateDriverRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - JWT: []
      summary: Update driver profile
      tags:
      - drivers
  /drivers/{id}/update:
    put:
      consumes:
      - application/json
      description: Updates a driver's specific field (e.g., VehicleInfo, CurrentLocation)
        based on driver ID
      parameters:
      - description: Driver ID
        in: path
        name: id
        required: true
        type: string
      - description: Field and value to update
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/driver.UpdateDriverRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid driver ID or request body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - JWT: []
      summary: Update a specific driver field
      tags:
      - drivers
  /drivers/all_drivers:
    get:
      description: Get a list of all registered drivers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/driver.Driver'
            type: array
      security:
      - JWT: []
      summary: List all drivers
      tags:
      - drivers
  /drivers/by-email/{email}:
    get:
      description: Retrieve a driver by their Email
      parameters:
      - description: Driver Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/driver.Driver'
        "400":
          description: Invalid Email
          schema:
            type: string
        "404":
          description: Driver not found
          schema:
            type: string
      security:
      - JWT: []
      summary: Get driver by Email
      tags:
      - drivers
  /drivers/by-id/{id}:
    get:
      description: Retrieve a driver by their ID
      parameters:
      - description: Driver ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/driver.Driver'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Driver not found
          schema:
            type: string
      security:
      - JWT: []
      summary: Get driver by ID
      tags:
      - drivers
  /drivers/create:
    post:
      consumes:
      - application/json
      description: Register a new driver with name, email, etc.
      parameters:
      - description: User Input
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/driver.CreateDriverRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/driver.Driver'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Failed to create driver
          schema:
            type: string
      security:
      - JWT: []
      summary: Create a new driver
      tags:
      - drivers
  /feedbacks/{id}:
    get:
      description: Retrieve a feedback by their ID
      parameters:
      - description: Feedback ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/feedback.Feedback'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Feedback not found
          schema:
            type: string
      security:
      - JWT: []
      summary: Get feedback by ID
      tags:
      - feedbacks
  /feedbacks/all_feedbacks:
    get:
      description: Get a list of all feedbacks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/feedback.Feedback'
            type: array
      security:
      - JWT: []
      summary: List all feedbacks
      tags:
      - feedbacks
  /feedbacks/create:
    post:
      consumes:
      - application/json
      description: Create a new feedback with order_id, customer_id, etc.
      parameters:
      - description: User Input
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/feedback.CreateFeedbackRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/feedback.Feedback'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Failed to create feedback
          schema:
            type: string
      security:
      - JWT: []
      summary: Create a new feedback
      tags:
      - feedbacks
...
schemes:
- http
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"